buckend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
faiss-cpu
sentence-transformers
